refactoring package name customExceptions to exceptions #9
